fixex in the code base

+++ src/utils/cloudinary.js
@@ -1,6 +1,5 @@
 import {v2 as cloudinary} from 'cloudinary';
 import fs from 'fs';
-import dotenv from 'dotenv';
 
 cloudinary.config({ 
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
@@ -20,7 +19,7 @@ const uploadToCloudinary=async (localfilepath)=>{
     } 
     catch (error) 
     {
-        fs.unlinkSync(localfilepath);
+        if(localfilepath) fs.unlinkSync(localfilepath);
         //remove the locally saved file if operation got failed
         return null;
     }
